move router logic into userControllers from authroutes.js, beginning, registerUser tested working
